refactor(bookings): move itemization editor from detail card to Payout page
The per-event price itemization now lives on the dedicated Booking
Payout route (/bands/{band}/booking/{booking}/payout), not in the
small Payout summary card on the booking detail page. The full Payout
screen is the more natural place to do financial breakdown work, and
the detail-card was getting cramped.

BookingPayout (the detail-page card) reverts to read-only summary.
BookingDetails goes back to v-if="payoutConfig" for the card.
Payout.vue (the standalone page) renders the itemization card just
below the header for multi-event bookings, with the same inline
on-blur save flow.

Dusk tests updated to visit the payout route instead of the detail
page when verifying itemization rendering and save behavior.
